[FIX] website_sale: align domain from route and act window
Act window domain are used to load tree view data when clicking on dashboard
metric.
That metric is retrieved through domain in backend.

This domain part wasn't adapted in act window domain (can't use ref).
As searching on website is enough to get ecommerce order, this line can be
safely removed, it is an old code that was needed before website_id was set on
order, as it can be noticed in 40a24d81a31.

X-original-commit: b31abe06583c06dfc8b0df047d2659e657785b89