[FIX] stock: apply push rule after transparent rule
Usecase to reproduce:
- Rule from WH/A -> WH/B as automatique no step added
- Rule from WH/B -> WH/C as transfer created

Do a move that goes to WH/A and confirm it. It's destination location
is correctly set to WH/B but the internal transfer from WH/B to WH/C
is not created.

It's due to the avoid loop location that returns always true due to
code removed in commit 68710b89e00d14f0babc55fdffce9c1ca7a650cc

closes odoo/odoo#61488

X-original-commit: b2a07e4d08beff360ac1e6c95d240a15b1081855
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>