[FIX] registry: unit test for #60899
The field `display_name` is a magic field automatically added to all
models.  On a custom model, it depends on `x_name`.

As it is automatically added by the ORM, `display_name` is considered a
`base` field, while `x_name` is considered a `manual` field.

On loading the transitive dependencies of `display_name` of a custom
model, for which the loading of the `x_name` has been skipped because it
depends on field not yet loaded, the exception was not ignored because
exceptions are ignored only for `manual` field, and `display_name` is
considered a `base` field.

Upgrade request 56274

X-original-commit: aff85a89ffae5268e412dbb0d390330908336c72
Co-authored-by: Raphael Collet <rco@odoo.com>